# Optimize fps, grayscale, trim