--- a/augly/video/augmenters/ffmpeg/fps.py
+++ b/augly/video/augmenters/ffmpeg/fps.py
@@ -19,8 +19,8 @@ class VideoAugmenterByFPSChange(BaseVidgearFFMPEGAugmenter):
 
         @param output_path: the path in which the resulting video will be stored.
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         command = [
             "-y",