--- a/augly/video/augmenters/ffmpeg/grayscale.py
+++ b/augly/video/augmenters/ffmpeg/grayscale.py
@@ -15,8 +15,8 @@ class VideoAugmenterByGrayscale(BaseVidgearFFMPEGAugmenter):
 
         @param output_path: the path in which the resulting video will be stored.
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         command = [
             "-y",