--- a/augly/video/augmenters/ffmpeg/grayscale.py
+++ b/augly/video/augmenters/ffmpeg/grayscale.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # Copyright (c) Facebook, Inc. and its affiliates.
 
-from typing import List, Optional
+from typing import List
 
 from augly.video.augmenters.ffmpeg.base_augmenter import BaseVidgearFFMPEGAugmenter