--- a/augly/video/augmenters/ffmpeg/trim.py
+++ b/augly/video/augmenters/ffmpeg/trim.py
@@ -45,8 +45,8 @@ class VideoAugmenterByTrim(BaseVidgearFFMPEGAugmenter):
 
         @param output_path: the path in which the resulting video will be stored.
 
-        @returns: a list of strings of the FFMPEG command if it were to be written
-            in a command line
+        @returns: a list of strings containing the CLI FFMPEG command for
+            the augmentation
         """
         video_info = get_video_info(video_path)
         duration = float(video_info["duration"])